Create nice helper for working with Maps (mapGetOrSet)
Also move some generic utils out of the composition utils file
and into their own tiny files in a utilities folder. This more
closely matches the structure of the gateway's code as well.
