+++ federation-js/src/utilities/isNotNullOrUndefined.ts
@@ -1,0 +1,5 @@
+export function isNotNullOrUndefined<T>(
+  value: T | null | undefined,
+): value is T {
+  return value !== null && typeof value !== 'undefined';
+}
